Move checklist_t to separate file
